pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "openai>=1.70.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "streamlit>=1.44.0",
    "xgboost>=3.0.0",
]
